Chroma from Luma (CfL) for Intra in PVQ
Uses the reconstructed Luma pixel as an intra prediction for the Chroma
pixels. In this first version, CfL is always on, which requires the
UV_DC_PRED_ONLY experiment so that Chroma intra prediction is not
signaled in the bitstream.

This preliminary version of CfL only supports 4:2:0 and does not perform CfL
during RDO. PVQ's gain prediction is used (which will be disabled in a
subsequent version). CfL flipping is also not implemented yet.

The CFL_CTX struct is introduced, it stores a pixel buffer per thread group.
CFL_CTX will replace PVQ's cfl_ctx struct in a subsequent version.

subset-1:
NO_Pix_CfL@2017-02-22T16:03:35.757Z -> Pix_CfL@2017-02-22T16:04:12.652Z
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
13.2282 |  6.7930 | 10.1853 |  13.6920 | 13.2804 | 13.1846 |    12.4776

Change-Id: Iac8c755c37119c818d60ab33592517d6d9389fa8