Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Was the warning like this?
> av1/decoder/decodeframe.c:1653:9: warning: implicit declaration of function ‘av1_decode_palette_tokens’ [-Wimplicit-function-declaration]
>          av1_decode_palette_tokens(xd, plane, r);
>          ^

Yes, I just replied the same on your patch :)
The thing is that when PVQ is enabled, palette tokens are never read in the decoder.
So, it seemed to me that the real fix for this is non-trivial, and you might want to think clearly about that.

But if the fix is clear to you, feel free to fix it for real right now.
I'll hold on this patch.

Patch-set: 1
Reviewer: Gerrit User 5290 <5290@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>