bitstream: void instruction for mi_row and mi_col
In the bitstream.c functions pack_inter_mode_mvs and
write_mb_modes_kf the arguments mi_row and mi_col are only used if
CONFIG_DELTA_Q set, or CONFIG_EC_MULTISYMBOL is not and CONFIG_CB4X4
is. This yields an unused parmeter warning if the required flags are
not set.

This patch adds a void instruction for each parameter in the case
that they aren't used.

Change-Id: I29bf1e33692b7d4c636d65fca9d130be9313fc9d