New boosting algorithm for xiphrc
This brings the CQ mode of the xiphrc experimental rate control system
closer to that of the current rate control system. This mainly affects
the metrics rather than visual performance.
It should be noted that the current rate control system does very big
boosts on keyframes/goldenframes which while increasing the overall
metrics throughout does affect visual quality in a negative way, and
pursuing matching metrics might not be a good idea.

Change-Id: I04ec65f9950a1f099a14016779806d058d46d1d0