Remove is_keyframe from pvq_decode_partition
Removes the pli and keyframe parameters and replaces them with
cfl_enabled. Since there's no H/V pred in AV1, Luma Intra is the same as
Luma Inter as far as PVQ is concerned. PVQ is only concerned whether CfL
is used or not.

Is it possible to review if all the conditions are required for CfL or
if some of them are related to H/V Pred.

There is no impact on the bitstream, as is_keyframe remains hardcoded to
0.

Change-Id: Idfa81acd9115dbca4b076ba7114c9ba88d0e44b1