Enable DC RDO Quant for keyframes in PVQ
When enabling is_keyframe, noticeable artifacts arise from the statement
that zeroes out the DC. This statement must be removed.

As for RDO quant of the DC, if we disable it for keyframes, we get the
following on subset1:
pvq@2017-03-01T13:48:01.310Z
  -> pvq-Keyframe-No-RDOQuantDC@2017-03-01T13:46:27.948Z
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.2651 | -2.9632 | -3.2333 |  -0.0073 | 0.2803 |  0.0719 |    -0.8823

If we enable RDO quant of the DC on keyframe only for the Luma plane, we
get the following on subset1:
pvq@2017-03-01T13:48:01.310Z
  -> pvq-Keyframe-Luma-RDOQuantDC@2017-03-01T13:51:46.886Z
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.2119 | -3.0653 | -3.4263 |   0.2075 | 0.2489 |  0.2082 |    -0.9186

If we enable RDO quant of the DC on keyframe only for the Chroma planes, we
get the following on subset1:
pvq@2017-03-01T13:48:01.310Z
  -> pvq-Keyframe-Chroma-RDOQuantDC@2017-03-01T13:53:43.026Z
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
0.0440 | -0.0233 | -0.0030 |  -0.2398 | 0.0191 | -0.1789 |     0.0322

The biggest impact arises from disabling RDO Quant for the DC of Chroma
plane on Intra frames. However, this still leads to a small regression
in Luma.

Change-Id: I61b6609c3c3a54e438b76f25ecd28e0b46e80b0b