Compute multiple global motion candidates and pick the best by error advantage.
This CL modifies ransac to keep the best N global motions by num_inliers and variance rather than a single one. The error advantage for each of the motions is computed in encode_frame_internal(), and the best one is used as the global motion for that pair of <target, reference> frames.

With N = 8, the improvement of global motion over the baseline on the lowres video set increases from -1.287 to -1.419.

Change-Id: Ic0c9066a3f175a5ea0a78828cd244104e70144ba