Replace is_keyframe checks with is_skip_copy check in PVQ
Because of H\V pred and CfL, block level skip on keyframes in Daala
implied zeroing out the coeff instead of copying the prediction. In AV1,
block level skips in keyframes impli copying the prediction.

To better reflect the nature of AV1, the checks are done on
is_skip_copy instead of is_keyframe. This will allow a skip zero
behavior if needed in the future.

is_keyframe is now enabled on keyframes for PVQ the changes are 0.0000
on AWCY. Enabling keyframe mostly changes the probabilities of keyframes
for EC. I'm unable to test with objective-fast at the moment because we
run out of RAM on AWCY.

Change-Id: I422f5a766ef4eb366e96df7a9068c8bd2f87666a