Add new compound coding modes for TRIPRED
This patch in particular is targeted to try new compound coding modes
for the single ref TRIPRED case, where two inter predictors are
obtained from the same reference frame for the inter-coded blocks.

Note that we may separate this from TRIPRED to a new coding tool, which
in essence proposes a new macroblock compound mode using a single
reference frame.

Change-Id: I3900973b96827d2a84226af1c478c10df252cf70