Replace is_keyframe with is_skip_copy in PVQ encoder
Because of H\V pred and CfL, block level skip on keyframes in Daala
implied zeroing out the coeff instead of copying the prediction. In
AV1, block level skips in keyframes only impli copying the prediction.

To better reflect the nature of AV1, the checks are done on
is_skip_copy instead of is_keyframe. This will allow a skip zero
behavior if needed in the future.

is_keyframe is now enabled on keyframes for PVQ the changes are 0.0000
(every metric) on AWCY. Enabling keyframe mostly changes the
probabilities of keyframes for EC. I'm unable to test with
objective-fast at the moment because we run out of RAM on AWCY.

Change-Id: I5d9ce77211c1df3fa6c11a30f097698d9d700c13