daala_ec: Remove non-dyadic functions.
Encoder output should not change, and all streams should remain
decodable without decoder changes.

The Laplace coder (as used by the generic coder) did have to change
because it truncated its Q15 CDFs when an upper-bound on the symbol
value was known, but this only happens in PVQ when disabling
"nodesync" for the theta values, which is not the default
configuration.

The truncation now uses the same Stuiver and Moffat technique as
the old non-dyadic coder to convert the truncated CDF back to Q15,
which should have minimal impact on BD-rate if it's ever enabled.

Change-Id: Id1f1b0f2f02c3b46f150a93c451bf48abd0782ca