Make encoder use vectorized self-guided filter
By rearranging the code in restoration.c, we can allow the
encoder to use the SSE4.1 version of the self-guided filter
while picking the loop-restoration filter.

This also helps us prepare for adding a highbitdepth SSE4.1
version of the self-guided filter.

No effect on encoder output, but gives an end-to-end speedup
of 1-2%.

Change-Id: Id17ba4a0963ddce9f70a7cae666e212e138d5f2c