Update patch set 15
Patch Set 15:

> Patch Set 15:
> 
> (2 comments)
> 
> > Also: Please clang-format analyzer.cc.
> 
> clang-format passes the style check because I have // clang-format off at the top of the CC file.
> 
> These style errors are coming from the linter, and in some cases are impossible to fix, e.g.,
> 
> examples/analyzer.cc:365:  Is this a non-const reference? If so, make const or use a pointer: wxCommandEvent &event  [runtime/references] [2]
> 
> I'm happy to fix some of the spacing issues, but this will never be clean.

clang-format off for the whole file seems unnecessary. Nothing is jumping out at me that clang-format is going to mangle. Out of curiosity: Why is it in here?

Please mark the wxWidgets API stuff that's angering the linter with NOLINTs. I hadn't noticed that it was more than just whitespace complaints.

Patch-set: 15