PVQ_CFL Experiment flag
This experiment does NOT replace the Chroma intra prediction
with the Luma reconstructed values (i.e. it does not perform CfL). It
only enables the PVQ mecanisms for it. PVQ is mislead to believe the
content is CfL in order to validate that it behaves properly.

OD_DISABLE_CFL is removed as CFL can be disabled by disabling the
PVQ_CFL flag. Enabling PVQ_CFL flag enables quantization on keyframe
DCs for PVQ. When a Chroma plane is skipped, the reference is not
copied, but the coefficients are zeroed.

CfL flipping is disabled as this feature can flip a value multiple
time during RDO.

Change-Id: I19621978bedb7ab4df6eb8050ad7781d35b3bce3