[PVQ_CFL] Disable uv_mode signaling
We disable uv_mode signaling in the bitstream and set
uv_mode = DC_PRED.

In other words, spatial Intra prediction is used to predict the
Chroma DCs and the ACs are all zeros. In a subsequent patch, the ACs
will be predicted using the reconstructed Luma pixels (CfL).

The computation of probabilities for EC is not modified.

Change-Id: I9ea301512bb4966f2ad886183e9c2515189e5274