[PVQ_CFL] Disable uv_mode signaling
We disable uv_mode signaling in the bitstream and set
uv_mode = DC_PRED.

In other words, spatial Intra prediction is used to predict the
Chroma DCs and the ACs are all zeros. In a subsequent patch, the ACs
will be predicted using the reconstructed Luma pixels (CfL).

Note that the computation of probabilities for EC is not modified.

Subset1:
pvq@2017-03-09T19:17:07.051Z
  -> DisableUVMode-7943@2017-03-09T19:18:16.429Z
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
3.1683 |  1.6431 |  2.7514 |   3.0238 | 3.1924 |  3.0864 |     2.8269

Change-Id: I9ea301512bb4966f2ad886183e9c2515189e5274