Chroma from Luma (CfL) for Intra in PVQ
Uses the reconstructed Luma pixel as an intra prediction for the
Chroma pixels. In this first version, CfL is always on.

This preliminary version of CfL only supports 4:2:0 and does not
perform CfL during RDO. PVQ's gain prediction is used and CfLà
flipping is disable to avoid RDO issues.

The CFL_CTX struct is introduced, it stores a pixel buffer per thread
group. CFL_CTX will replace PVQ's cfl_ctx struct in a subsequent
version.

When 4:2:0 is used, it is possible to have Chromatic overrun (i.e.
when the surface covered by Chroma is bigger than the surface covered
by luma). When this happens we pad by copying the last available luma
pixel value.

Change-Id: I20e27508aab493dacc6ebc0aa8082d6d53de9580