SMOOTH_PRED: Use a table to get log2 of block dimension.
Using a table is feasible because the max block dimension is bounded.

Also, using the floating point operation log2() was resulting in an
assert failure due to an unrelated floating point exception that
happens earlier.

Related: update the MD5s in test_intra_pred_speed to fix that failure
too.

BUG=aomedia:384

Change-Id: I18dc0733e880bac21b3d07ad874f8ae341f59f06