handle_inter_mode: Motion mode RD calc to function
Refactors the end of handle_inter_mode into a new funciton. This code
is responsible for calculating an accurate RD for the SIMPLE_TRANSLATION
motion mode in the simplest case, and does the same for other motion
modes as their experiments are enabled.

This patch aims to do as little as possible to the code inside the
function - that is left to later patches to reduce the complexity of
this diff.

Change-Id: I62bf5aae34594b0a1dc4813aeba99e675d6db374