Macro to disable warped/obmc_causal with global motion
Enabling SEPARATE_GLOBAL_MOTION will remove the ability for
a block that uses zeromv with global motion to pick warped_causal
or obmc_causal as the motion mode. When this is enabled there is:

0.05% drop on lowres for global + warped enabled
0.153% drop on midres for global + warped enabled
No performance change to global, warped, or motion var individually.

Still waiting for final results after the rebase

Change-Id: Idbfb8dd7a93da14902438504b06a08e5212e48cb