[PVQ_CFL] Merge CFL_CTX with cfl_ctx
In Daala, cfl_ctx was a decoder side struct for the CfL context. In AV1,
the context struct CFL_CTX is shared by both the encoder and decoder.
Merging CFL_CTX with cfl_ctx requires that more CfL related code in PVQ
be moved into the PVQ_CFL experiment.

This change does not have an impact on the bitstream.

Change-Id: I945be3507ea902eb8c95a10f27006d40df1c6cf8