[PVQ_CFL] AC Coeff Flip Bit
Because the Chroma coefficients are sometimes inversely correlated with the
coincident luma coefficients, the sign of the AC coefficients of the
predictor must be flipped.

A single bit is signaled in the bitstream to indicate whether to flip
the sign of the ACs. The flip bit is signaled before the first with a
reference (noref=0)

Subset1:
pvq@2017-03-15T20:26:14.990Z -> CfL-8566@2017-03-15T20:25:02.606Z
  PSNR | PSNR Cb | PSNR Cr | PSNR HVS |   SSIM | MS SSIM | CIEDE 2000
1.9717 | -7.2514 | -5.1211 |   1.9462 | 2.0124 |  1.9740 |    -1.4424

Change-Id: I038773c9eea7e22c4a1115ce5ff00c43d3bd338e