unify the intra prediction angle options for different block sizes
this change unifies the intra prediction angles so that 8x8 block has
the same number of angles and angle steps as 16x16 and 32x32 blocks.
the benefit is that HW implementation can have the same logic to
handle all block sizes. otherwise, prediction angles for 8x8 is
different from other block sizes, which requies special logic.

Change-Id: I0889e8f859f60447f0b505ac743278c33d2bff13