Replace division in warped motion least squares
Replaces the int64 division in least-squares with a
mechanism that decomposes the divisor D such that
1/D = y * 2^-k where y is obtained from a lookup
table. The main complexity is now from computing
this decomposition, which is essentially equivalent
to finding floor(log2(x)) (position of highest
bit in a 64-bit integer) using O(logn) operations.

Also includes an out of memory bug fix and some cleanups.

Change-Id: I9247fdff5f6b4191175d4b4656357bfff626f02c