Update patch set 1
Patch Set 1:

> I'm not sure what you're trying to do exactly, but I'm almost
 > certain it's incorrect.

I'd like to see what AWCY says if I feed the unfiltered (not deringed) frame to CLPF, calculate the delta, and add that delta to the deringed frame.  It's a shot in the dark, but it would break the dependency between dering and clpf.

This patch still has a bug.  I have to disable filtering across the border below for every 8x8 block to avoid a enc/dec mismatch.  I'm suspecting it could have something to do with the bottom edge of the frame which doesn't line up with a sb boundary (perhaps the frame isn't padded both in the encoder and decoder?)

Let's discuss this in the talk today.

Patch-set: 1
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>