Update patch set 2
Patch Set 2:

> Can you move the boundary removal to a separate patch? That will
 > make it easier to test both separately. Otherwise, the patch looks
 > good to me. We'll just need to fix the chroma regression.

I already did that as a separate patch #8655.  The only thing it missed was that od_dering still took boundary_info as an (unused) argument, which is amended in this patch (assuming that we want dering() to handle boundaries, not od_dering()).  It's slightly cleaner if that detail was moved to #8655, but I'm not sure if I'm able to do that cleanly in git since there will be merge hell.

Patch-set: 2
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>