Update patch set 3
Patch Set 3:

> Looks good. I'll +2 it when the tests pass.

The failure is an unused argument in od_dering (not really an error...), which is fixed in the other patch.  Fixing it might trigger a merge issue.  Are you ok to +2 both and then we can submit both together?  The other patch passes the tests and is a child of this patch.

Patch-set: 3
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>