Re-enable av1_nmv_ctx() when ext-inter is enabled
Currently, mbmi->ref_mv_idx can be set to a nonzero value
on the encoder side if mbmi->mode is one of NEARMV, NEWMV,
NEAR_NEARMV, or NEW_NEWMV. But it can only be nonzero on the
decoder side if the mode is NEARMV or NEWMV. Hence av1_nmv_ctx
has previously been disabled when ext-inter is enabled, to
prevent a mismatch due to this.

This patch changes the encoder behaviour to match the decoder
behaviour.

Change-Id: Icfe41fb72e76731ae373fe8c6065f5e003f6414f