frame_superres: Skeleton code and todos
This code isn't included anywhere yet. My goal is for the current set of
functions to be the only ones exposed to the rest of the code base -
this should make things simple. A pre-decode function may be needed
later. The new superres files have been added to the cmake files so they
are a part of the build now.

Copyright info has also been added.

Change-Id: I5b74b78d654ae73943c99c7d3312a44103d00bde