WIP: frame_superres: Post encode and decode upscaling
This is a work in progress. It is not ready for full review or
submission.

This patch implements the post-encode and post-decode upscaling for the
frame superresolution experiment to work. Pre-encode downscaling was
already working but was disabled by never setting the superres numerator
to anything other than zero.

Upscaling happens after cdef and before loop restoration.

For now, this patch forces on random-superres. This should to be
disabled before submission.

Change-Id: If0a8f69224dfaa0f4ae7703bd429ea2af953c7a6