Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)

This file contains changes that are not compatible with the version of the codec when this experiment is switched off. What are your suggestions to overcome this problem? 
 
A suggestion from me: Since this experiment has already been adopted at the last f-2-f meeting, proceed with the source code adoption, clean all #if statements from the code and remove this include (it will not be needed after that).

Patch-set: 1
Reviewer: Gerrit User 6720 <6720@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>