Simplify warped motion estimation to use 2d ls
Use a simpler warped motion estimation scheme that uses a 2d
least squares problem, where the underlying assumption
applied is that the motion vector computed at the center
of the current block using the warp model is exactly the same
as the motion vector transmitted for the block.

The main motivation is to reduce the complexity of the
estimation process.

Coding efficiency drop is about +0.25% on lowres:
-1.152% (from -1.396%).

Also, removes code for non-approximate division and bakes
approximate divison in.

Change-Id: Ie4ad8e32593b09f7e1920c70b0b92545236ddc54