Add codes for coding based on reference
Adds a binary code for coding symbols from an n-ary
alphabet based on a reference value for the symbol.
The code has two steps. If the symbol is close to
the reference a shorter code is used, while if it is
farther away a longer code is used.

This code will be used for more efficient encoding of
loop-restoration parameters and global motion
parameters.

Change-Id: I28c82b611925c1ab17f544c48c4b1287930764b7