Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Does any encoder change require this md5 change?

to the baseline it would. Jim was using this as a canary to flag a potential bitstream change (which at one point I thought were supposed to be tagged with [BITSTREAM]). it's extremely fragile. another option might be to pin a decoder version and try it against some encodes produced by the patch

Patch-set: 2
CC: Gerrit User 5020 <5020@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>