Make CLPF correctly handle boundaries
8 bit CLPF doesn't handle borders for values near 255, but 16 bit CLPF
does.  So we can use 16 bit CLPF near these borders then convert to 8
bit.  This is slightly suboptimal, but the vast majority of blocks can
use 8 bit CLPF.

Perf reports the following when decoding 600 frames of Crowdrun (8 bit):

   1.32%  aomdec  aomdec  [.] clpf_block8
   1.21%  aomdec  aomdec  [.] clpf_block4
   0.24%  aomdec  aomdec  [.] clpf_hblock8
   0.07%  aomdec  aomdec  [.] aom_clpf_block_sse4_1
   0.07%  aomdec  aomdec  [.] clpf_block_hbd
   0.05%  aomdec  aomdec  [.] clpf_block_hbd4
   0.01%  aomdec  aomdec  [.] clpf_hblock_hbd
   0.01%  aomdec  aomdec  [.] aom_clpf_hblock_sse4_1
   0.01%  aomdec  aomdec  [.] clpf_hblock4

Only 0.13% of the time is spent in the 16 bit versions, but these are
still doing the filter, so the overhead is less.  In addition there is
an extra 16 to 8 bit conversion, but once those are simdified, the
cost of this solution is likely < 0.1% and quite acceptable.

Change-Id: I9612861ce3026f8527c166999b047a62a94628f4