Update patch set 3
Patch Set 3:

Should not be merged before we agree that this approach is what we want.  I haven't yet compared the performance with the other option which is to go fully 16 bit only for the edges.

Note that the packing was changed from signed->unsigned to signed->signed.  This is ok since the zero point for the clpf damping is at +/-2^6 at the most.  I don't think that will change, so going to signed->signed should be ok.  That makes it possible to do abs and to fetch the sign after the packing, which saves many instructions.

Patch-set: 3
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>