Update patch set 3
Patch Set 3:

As far as I can tell, the new version is should be using 1 fewer instruction than the previous version, so it should be at least as fast for 8-bit as what we have now. From there, removing the handling of the padding would only save one instruction -- not sure it's worth it. And even if it was really worth it, I'd rather have the function duplicated than the solution from 9120.

Patch-set: 3
Reviewer: Gerrit User 5045 <5045@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>