Update patch set 4
Patch Set 4:

The failing test is the test/decode_to_md5.sh script. It's a new test that checks for a bit-exact output. Of course that fails every time we try to improve, so we have to change the md5 checksum in the script but nobody's been able to tell me how to do that.

Patch-set: 4
Reviewer: Gerrit User 5045 <5045@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>