Update patch set 6
Patch Set 6:

This is a normative change.  Instead of clipping to 0 when the dering difference is above the threshold, use the CLPF constrain function instead.  The best "damping" values were found to be 6 for luma and 5 for chroma (whereas CLPF prefers something qp dependent).

The constrain_hbd was renamed constrain16 and put in cdef_simd.h so it can be shared between the dering and clpf code.

I'm getting cpu-used=0 numbers as well.  The total effect of CDEF for ll4 (without JM's joint chroma search) is:

   PSNR|PSNR Cb|PSNR Cr|PSNR HVS|   SSIM|MS SSIM|CIEDE 2000
-7.7994|-4.0015|-4.0660| -5.5117|-8.4922|-6.5618|   -7.2504

Patch-set: 6
Reviewer: Gerrit User 5111 <5111@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>