Update patch set 4
Patch Set 4:

I've been running tests on low-latency, cpu-used=4 and it seems like cdef is losing some gain when cb4x4 is enabled. 

Without cb4x4, the bd-rate for cdef-off vs cdef-on is:
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
-7.0054 | -4.4064 | -4.5621 |  -5.1048 | -7.9425 | -6.0554 |    -6.8543

When cb4x4 is enabled, then cdef-off vs cdef-on is:
   PSNR | PSNR Cb | PSNR Cr | PSNR HVS |    SSIM | MS SSIM | CIEDE 2000
-6.2863 | -4.0866 | -4.3063 |  -4.5984 | -7.3339 | -5.5651 |    -6.2671

So we're losing about 0.7% PSNR. It doesn't necessarily mean a bug, but it's worth investigating.

Patch-set: 4
CC: Gerrit User 5045 <5045@e5514cf8-2d6e-3e29-adb4-24cd6dde4bf0>