ext-inter + ref-mv: Let NEW_NEARMV and NEAR_NEWMV use ref_mv_idx
This means that all compound modes with any NEW or NEAR components
can use the drl system to indicate which reference mvs to use.

Note: There is a slight overlap between (say) NEW_NEARMV and
NEW_NEARESTMV, in that they code the same motion vector when
ref_mv_idx = 0. But changing to avoid this case seems to
remove much of the gain from this patch, so we leave this
case in for now.

Change-Id: Iaa7d7a46ca2e33e1f05a214772c4f3f937d14e5f