frame_superres: Add normative scale and extend frame
This commit adds a normative version of scale and extend frame to the
now common set of resizing funcitons. This is just a copy of the
nonnormative scale and extend frame function, but it can now be used in
a normative manner since it is in common and only depends on other code
in common.

Change-Id: I83afc3d97ff8ebce5ecb2e1205ba6c271156cb31